Changed Piece draw, added Drawable interface, changed how textures are loader, added 3 new interface buttons and a Spawner that has a distance heuristic function to dertermine the fastest path to the exit.

+++ projects/WarInTheEast/src/Camera.cpp
@@ -70,12 +70,12 @@ void Camera::setCenter(glm::vec3 v)
 	updateCamera();
 }
 
-void Camera::setLast_mx(int x)
+void Camera::setLast_mx(float x)
 {
 	last_mx = x;
 }
 
-void Camera::setLast_my(int y)
+void Camera::setLast_my(float y)
 {
 	last_my = y;
 }
